+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project changes will be shown here.
 ## [unreleased]
 ### Added
 - Issue 951 - Expenses Breakdown Year to Date
+- Issue 890 - Monitor Sales Order to get notifications when it's signed.
 ### Changed
 - Issue 938 - Increase number of favourites in menu
 - Issue 950 - Include team in time allocation